requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
google-generativeai==0.3.0
textblob==0.17.1
langchain-huggingface==0.0.29
langchain-core==0.1.28
transformers==4.35.2
pandas==2.1.3
pillow==10.1.0
huggingface-hub==0.19.4
httpx==0.25.2
pydub==0.25.1
vosk==0.3.45
requests==2.31.0
torch==2.1.1
